Python-Backend/requirements.txt
pdfplumber==0.10.3
python-docx==1.1.0
fastapi==0.110.0
uvicorn==0.29.0
numpy==1.26.4
faiss-cpu==1.7.4
sentence-transformers==2.6.1
torch==2.2.2
transformers==4.38.2
python-dotenv
groq
pandas
